requirements.txt
wheel==0.45.1
processcube-client==6.0.5
#docker==6.1.3
docker==7.1.0
requests>=2.31.0
robotframework>=7.3.2
six==1.17.0
